fix: fixed double to int conversion

+++ src/Figure.java
@@ -1,6 +1,5 @@
 public abstract class Figure implements ICalculate {
     public double area;
-    public int perimeter;
-
+    public double perimeter;
     public abstract void getPerimeter();
 }
